moved initialState to default
